requirements.txt
beautifulsoup4==4.12.3
facenet_pytorch==2.6.0
fastapi==0.111.1
htmldate==1.8.1
Katna==0.9.2
langchain_openai==0.1.17
matplotlib==3.9.1
newspaper==0.1.0.7
numpy==2.0.0
openai==1.36.0
pandas==2.2.2
Pillow==9.0.1
Pillow==10.4.0
pinecone==4.0.0
protobuf==3.12.4
pydantic==1.10.15
pytube==15.0.0
Requests==2.32.3
serpapi==0.1.5
streamlit==1.36.0
tensorflow==2.17.0
torch==2.3.1
torchvision==0.18.1
transformers==4.42.3
uvicorn==0.30.3
